requirements.txt
django>=5.2.1
tensorflow>=2.15.0
pillow>=10.2.0
whitenoise>=6.6.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
gunicorn==21.2.0 